% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fratio2D.R
\name{fratio2D}
\alias{fratio2D}
\title{"fratio2D" function for MCMC posterior sampling}
\usage{
fratio2D(t1, t2, myData, choice)
}
\arguments{
\item{t1}{An array containing 5*1 parameters}

\item{t2}{An array containing another 5*1 parameter}

\item{myData}{3D array containing 2 dimensional landmark}

\item{choice}{Which to compare with 1st object or mean obj}
}
\value{
fratio value
}
\description{
2D landmark data(Gaussian likelihood with Isotropic Error Variance)
as the name suggest, it evaluates fratio for two parameter vector t1, t2
}
\details{
"c1","c2" = Location parameter
"b" = dilation parameter
"theta" =Rotation angle
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
for(k in 1:10)
{tempp2=rnorm(5, mean=1, sd=1);
tempp1=purturb2D(tempp2,rep(0.001,5))
p=fratio2D (tempp1, tempp2,  myData, 10)
print(p)
}
}
}
\keyword{fratio2D}
