% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellotest.R
\name{hellotest}
\alias{hellotest}
\title{Hello, Testing Hello !
This is an example function named 'hello'
which prints 'Testing Hello!'.}
\usage{
hellotest()
}
\value{
testing hello !
}
\description{
Hello, Testing Hello !
This is an example function named 'hello'
which prints 'Testing Hello!'.
}
\examples{
\dontrun{
hellotest()
}

}
\keyword{hellotest}
