% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helmert.R
\name{Helmert}
\alias{Helmert}
\title{"Helmert"  computes The Helmert sub-matrix}
\usage{
Helmert(dim)
}
\arguments{
\item{dim}{dimension of Helmert matrix}
}
\value{
Returns The Helmert sub-matrix of dimension (dim - 1, dim)
}
\description{
"Helmert"  computes The Helmert sub-matrix
}
\examples{
\dontrun{
Helmert(2)
Helmert(3)
}
}
\keyword{Helmert}
