% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCpostPsample2D.R
\name{MCMCpostPsample2D}
\alias{MCMCpostPsample2D}
\title{MCMC posterior sampling for 2D landmark data (in Pre-shape space)
(Gaussian likelihood with Isotropic Error Variance)}
\usage{
MCMCpostPsample2D(initial, tune, myData, Nsample)
}
\arguments{
\item{initial}{The start value of  parameter sigma for MCMC run}

\item{tune}{Tuning value of MCMC sampler}

\item{myData}{3D array containing 2 dimensional landmark}

\item{Nsample}{Number of MCMC sample desired}
}
\value{
matrix containing samples from posterior density of parameter
}
\description{
Draws posterior from 5 parameter "Sigma" from Whole Data with pre-shape spaced landmarks
}
\details{
"Sigma" = Isotropic error variation parameter
Note that, here we are assuming Isotropy of error variance
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
for(i in 1:dim(myData)[3])
{myData[,,i]=
Morpho::rotonto(myData[,,1],myData[,,i])$yrot}
ape5000=MCMCpostPsample2D(1.5,
rep(0.1,1),myData,5000)
head(ape5000)
}

}
\keyword{MCMCpostPsample2D}
