% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfratio2D.R
\name{Pfratio2D}
\alias{Pfratio2D}
\title{"Pfratio2D" function for MCMC posterior sampling}
\usage{
Pfratio2D(t1, t2, myData)
}
\arguments{
\item{t1}{An array containing 1 parameter sigma}

\item{t2}{An array containing another 1 parameter sigma}

\item{myData}{3D array containing 2 dimensional landmark}
}
\value{
fratio value
}
\description{
2D landmark data in Pre-shape space (Gaussian likelihood with Isotropic Error Variance)
as the name suggest, it evaluates fratio for two parameter vector t1, t2
}
\examples{
\dontrun{
require(shapes);
require(Morpho);
data(apes)
myData = apes$x
for(i in 1:dim(myData)[3])
{myData[,,i]=Morpho::rotonto(myData[,,1],myData[,,i])$yrot
}
for(k in 1:10)
{tempp2=runif(1, 0.1, 2);
tempp1=Ppurturb2D(tempp2,rep(0.01,1))
p=Pfratio2D (tempp1, tempp2,  myData)
print(p)
}
}
}
\keyword{Pfratio2D}
