% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTpostvar2D.R
\name{PLOTpostvar2D}
\alias{PLOTpostvar2D}
\title{"PLOTpostvar2D"  Plot of posterior of Landmark variance parameter from MCMC sampling}
\usage{
PLOTpostvar2D(mcmcmat, burnin)
}
\arguments{
\item{mcmcmat}{A matrix containing MCMC sample values from posterior of  5*1 parameters}

\item{burnin}{Value for MCMC burn-in}

\item{Nsample}{number of MCMC samples}
}
\value{
plot
}
\description{
2D landmark data(Gaussian likelihood with Isotropic Error Variance)
Plot of posterior of Landmark variance parameter
}
\details{
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
ape10000=MCMCpostsample2D(rnorm(5, mean=1, sd=1),rep(1,5),
apes$x,10,10000)
PLOTpostvar2D(ape10000, 1000)

}
}
\keyword{PLOTpostvar2D}
