% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppurturb2D.R
\name{Ppurturb2D}
\alias{Ppurturb2D}
\title{"Ppurturb2D" function for Pre-Shape space Data & MCMC posterior sampling}
\usage{
Ppurturb2D(t, tune)
}
\arguments{
\item{t}{An array containing 5*1 parameters}

\item{tune}{array containing 5*1 tuning value}
}
\value{
g, which is a (new) purturbed point from parameter space
}
\description{
2D landmark -Pre-Shape space Data(Gaussian likelihood with Isotropic Error Variance)
generates pertrubed point from 1 parameter sigma.
}
\details{
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
NN=10; temp=rep(NA,NN); tempp=rep(1,1);
plot(x = 1,type = "n",xlim = c(0, NN+1), ylim = c(-10, 10),pch = 16,xlab = "steps", ylab = "mean(t)",main = "Purturbed point plot")
abline(h=mean(tempp))
for(k in 1:NN)
{   Sys.sleep(.1)
temp[k]=mean(Ppurturb2D(tempp, rep(1,1)))
 points( temp[k]~k,pch = 10,col =k)
Sys.sleep(0)
}
}
}
\keyword{Ppurturb2D}
