% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prun2D.R
\name{Prun2D}
\alias{Prun2D}
\title{"Prun2D"  MCMC sampling function for running MCMC sampler}
\usage{
Prun2D(t, tune, myData, Nsample)
}
\arguments{
\item{t}{An array containing 1 parameters}

\item{tune}{array containing 1 tuning value}

\item{myData}{3D array containing 2 dimensional landmark}

\item{Nsample}{number of MCMC sample to  generated}
}
\value{
Returns a large  matrix containing the actual samples from posterior of 1 parameter point
}
\description{
2D landmark data in Pre Shape  Space (Gaussian likelihood with Isotropic Error Variance)
Accepts new parameter 5*1 point with probability alpha
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
for(i in 1:dim(myData)[3])
{myData[,,i]=
Morpho::rotonto(myData[,,1],myData[,,i])$yrot}
r=Prun2D(runif(1, 0.1,1),rep(0.01,1),
 myData, 5000)
head(r); dim(r)
}
}
\keyword{Prun2D}
