% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pstep2D.R
\name{Pstep2D}
\alias{Pstep2D}
\title{"Pstep2D"  MCMC sampling function: Accept new point with probability alpha}
\usage{
Pstep2D(t, tune, myData)
}
\arguments{
\item{t}{An array containing 1 parameters}

\item{tune}{array containing 1 tuning value}

\item{myData}{3D array containing 2 dimensional landmark}
}
\value{
Returns the 1 parameter point
}
\description{
2D landmark data in Pre Shape space(Gaussian likelihood with Isotropic Error Variance)
Accepts new parameter 5*1 point with probability alpha
}
\details{
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
for(i in 1:dim(myData)[3])
{myData[,,i]=Morpho::rotonto(myData[,,1],myData[,,i])$yrot}
tt=rep(1,1)
for(i in 1:3){
s=Pstep2D(tt,rep(0.01,1),myData)
print(s)
tt=Ppurturb2D(tt, rep(0.01,1))
}
}
}
\keyword{Pstep2D}
