% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purturb2D.R
\name{purturb2D}
\alias{purturb2D}
\title{"purturb2D" function for MCMC posterior sampling}
\usage{
purturb2D(t, tune)
}
\arguments{
\item{t}{An array containing 5*1 parameters}

\item{tune}{array containing 5*1 tuning value}
}
\value{
g, which is a (new) purturbed point from parameter space
}
\description{
2D landmark data(Gaussian likelihood with Isotropic Error Variance)
generates pertrubed point from 5 parameter space.
}
\details{
"c1","c2" = Location parameter
"b" = dilation parameter
"theta" =Rotation angle
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
NN=10; temp=rep(NA,NN); tempp=rep(1,5);
plot(x = 1,type = "n",xlim = c(0, NN+1), ylim = c(-10, 10),pch = 16,xlab = "steps", ylab = "mean(t)",main = "Purturbed point plot")
abline(h=mean(tempp))
for(k in 1:NN)
{   Sys.sleep(.1)
temp[k]=mean(purturb(tempp, rep(11,5)))
 points( temp[k]~k,pch = 10,col =k)
Sys.sleep(0)
}
}
}
\keyword{purturb2D}
