% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2D.R
\name{run2D}
\alias{run2D}
\title{"run2D"  MCMC sampling function for running MCMC sampler}
\usage{
run2D(t, tune, myData, choice, Nsample)
}
\arguments{
\item{t}{An array containing 5*1 parameters}

\item{tune}{array containing 5*1 tuning value}

\item{myData}{3D array containing 2 dimensional landmark}

\item{choice}{Which to compare with 1st object or mean obj}

\item{Nsample}{number of MCMC sample to  generated}
}
\value{
Returns a large  matrix containing the actual samples from posterior of 5*1 parameter point
}
\description{
2D landmark data(Gaussian likelihood with Isotropic Error Variance)
Accepts new parameter 5*1 point with probability alpha
}
\details{
"c1","c2" = Location parameter
"b" = dilation parameter
"theta" =Rotation angle
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
r=run2D(rnorm(5, mean=1, sd=1),rep(0.1,5),
 myData, 10, 5000)
head(r); dim(r)
}
}
\keyword{run2D}
