% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2D.R
\name{step2D}
\alias{step2D}
\title{"step2D"  MCMC sampling function: Accept new point with probability alpha}
\usage{
step2D(t, tune, myData, choice)
}
\arguments{
\item{t}{An array containing 5*1 parameters}

\item{tune}{array containing 5*1 tuning value}

\item{myData}{3D array containing 2 dimensional landmark}

\item{choice}{Which to compare with 1st object or mean obj}
}
\value{
Returns the 5*1 parameter point
}
\description{
2D landmark data(Gaussian likelihood with Isotropic Error Variance)
Accepts new parameter 5*1 point with probability alpha
}
\details{
"c1","c2" = Location parameter
"b" = dilation parameter
"theta" =Rotation angle
"Sigma" = Isotropic error variation parameter
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
tt=rep(1,5)
for(i in 1:3){
s=step2D(tt,rep(0.1,5),myData,10)
print(s)
tt=purturb2D(tt, rep(0.1,5))
}
}
}
\keyword{step2D}
