% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMCMCpostsample2D.R
\name{TMCMCpostsample2D}
\alias{TMCMCpostsample2D}
\title{MCMC posterior sampling for 2D landmark data
(Gaussian likelihood with Isotropic Error Variance)}
\usage{
TMCMCpostsample2D(myData, choice, Nlandmark, tune, Nsample)
}
\arguments{
\item{myData}{3D array containing 2 dimensional landmark}

\item{choice}{Which to compare with 1st object or mean obj}

\item{Nlandmark}{How many landmarks are there for each object? (usually nrow(myData\link{,,1}))}

\item{tune}{Tuning value of MCMC sampler}

\item{Nsample}{Number of MCMC sample desired}
}
\value{
matrix containing samples from posterior density of parameter
}
\description{
Draws posterior from 5 parameters "c1","c2", "b", "theta", "Sigma"
}
\details{
"c1","c2" = Location parameter
"b" = dilation parameter
"theta" =Rotation angle
"Sigma" = Isotropic error variation parameter
Note that, here we are assuming Isotropy of error variance
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
ape_ress_10000=MCMCpostsample2D(apes$x,10,nrow(myData[,,1]), tune=1,10)
}

}
\keyword{TMCMCpostsample2D}
